chore: Update nx.json and tsconfig.json files, and add vitest.config.ts
This commit updates the nx.json file to simplify the "dependsOn" property in the "targetDefaults" section. It also removes the "cacheDirectory" property from the "options" object in the "release" section. Additionally, it removes the "test-1.spec.ts" file from the repository. Furthermore, it updates the tsconfig.json file to remove the "vitest/importMeta" type from the "types" array. Finally, it adds the vitest.config.ts file to the repository, which includes configuration for running unit tests with Vite and Playwright.