chore: Update npm dependency to latest stable version
This commit updates the npm dependency for the "msw" package to the latest stable version, which is 1.3.3. This update ensures that the project is using the most up-to-date features and bug fixes provided by the package.

These changes improve the stability and functionality of the project by incorporating the latest updates from the "msw" package.